Separerer ut validering av behandlingsresultat som eit eiget konsept i ei eiga fil
